Sort events by start time.
