+++ src/features/events/hooks/useAllEvents.ts
@@ -2,6 +2,7 @@ import { useApiClient, useAppSelector } from 'core/hooks';
 import useRemoteList from 'core/hooks/useRemoteList';
 import { allEventsLoad, allEventsLoaded } from '../store';
 import getAllEvents from '../rpc/getAllEvents';
+import sortEventsByStartTime from '../utils/sortEventsByStartTime';
 
 export default function useAllEvents() {
   const apiClient = useApiClient();
@@ -9,7 +10,10 @@ export default function useAllEvents() {
 
   return useRemoteList(allEventsList, {
     actionOnLoad: () => allEventsLoad(),
-    actionOnSuccess: (data) => allEventsLoaded(data),
+    actionOnSuccess: (data) => {
+      const sorted = data.sort(sortEventsByStartTime);
+      return allEventsLoaded(sorted);
+    },
     loader: () => apiClient.rpc(getAllEvents, {}),
   });
 }
